docblocks, site configuration file, user roles constants in the user model

+++ routes/web.php
@@ -16,7 +16,7 @@ Route::get('/setup', function() {
     return 'Initial admin setup!';
 })->name('setup');
 
-Route::get('/', 'DashboardController@index')->name('home');
+Route::get('/', 'DashboardController@index')->name('home')->middleware('auth');
 
 // Users resource
 Route::group(['prefix' => 'users', 'middleware' => 'auth', 'as' => 'users.'], function() {
@@ -25,33 +25,23 @@ Route::group(['prefix' => 'users', 'middleware' => 'auth', 'as' => 'users.'], fu
     Route::post('/create-admin', 'UsersController@storeAdmin')->middleware('owner');
 });
 
-// Objects resource
+// Objects object
 Route::group(['prefix' => 'objects', 'middleware' => 'auth', 'as' => 'objects.'], function() {
     Route::get('/', 'ObjectsController@index')->name('index');
 
-    // Create and store the resource
-    Route::get('/create', 'ObjectsController@create')->name('create');
-    Route::post('/', 'ObjectsController@store')->name('store');
+    // Create and store the object
+    Route::get('/create', 'ObjectsController@create')->name('create')->middleware('owner');
+    Route::post('/', 'ObjectsController@store')->name('store')->middleware('owner');
 
-    // Edit and update the resource
-    Route::get('/{id}', 'ObjectsController@edit')->name('edit')->where('id', '[0-9]+');
-    Route::put('/{id}', 'ObjectsController@update')->name('update')->where('id', '[0-9]+');
+    // Edit and update the object
+    Route::get('/{id}', 'ObjectsController@edit')->name('edit')->where('id', '[0-9]+')->middleware('owner');
+    Route::put('/{id}', 'ObjectsController@update')->name('update')->where('id', '[0-9]+')->middleware('owner');
 
-    // Destroy the resource
-    Route::delete('/{id}', 'ObjectsController@destroy')->name('destroy')->where('id', '[0-9]+');
+    // Destroy the object
+    Route::delete('/{id}', 'ObjectsController@destroy')->name('destroy')->where('id', '[0-9]+')->middleware('owner');
 });
 
 // Authentication Routes...
-Route::get('login', 'Auth\LoginController@showLoginForm')->name('login');
-Route::post('login', 'Auth\LoginController@login');
-Route::post('logout', 'Auth\LoginController@logout')->name('logout');
-
-// Registration Routes...
-//Route::get('register', 'Auth\RegisterController@showRegistrationForm')->name('register');
-//Route::post('register', 'Auth\RegisterController@register');
-
-// Password Reset Routes...
-//Route::get('password/reset', 'Auth\ForgotPasswordController@showLinkRequestForm');
-//Route::post('password/email', 'Auth\ForgotPasswordController@sendResetLinkEmail');
-//Route::get('password/reset/{token}', 'Auth\ResetPasswordController@showResetForm');
-//Route::post('password/reset', 'Auth\ResetPasswordController@reset');
+Route::get('login', 'LoginController@showLoginForm')->name('login');
+Route::post('login', 'LoginController@login');
+Route::post('logout', 'LoginController@logout')->name('logout');
